Cache LIL data lookups
