+++ capstone/capweb/helpers.py
@@ -1,10 +1,43 @@
 import json
+from functools import wraps
+
 import requests
 import django_hosts
+from django.conf import settings
+from django.core.cache import caches
 from django.urls import NoReverseMatch
 from django_hosts.resolvers import get_host_patterns
 
 
+def cache_func(key, timeout=None, cache_name='default'):
+    """
+        Decorator to cache decorated function's output according to a custom key.
+        `key` should be a lambda that takes the decorated function's arguments and returns a cache key.
+    """
+    cache = caches[cache_name]
+    def decorator(func):
+        @wraps(func)
+        def decorated(*args, **kwargs):
+            cache_key = key(*args, **kwargs)
+
+            # return existing value, if any
+            value = cache.get(cache_key)
+            if value is not None:
+                print("Got existing for", cache_key)
+                return value
+            print("Making new for", cache_key)
+
+            # cache new value
+            value = func(*args, **kwargs)
+            cache.set(cache_key, value, timeout)
+            return value
+        return decorated
+    return decorator
+
+@cache_func(
+    key=lambda section: 'get_data_from_lil_site:%s' % section,
+    timeout=settings.CACHED_COUNT_TIMEOUT
+)
 def get_data_from_lil_site(section="news"):
     response = requests.get("https://lil.law.harvard.edu/api/%s/caselaw-access-project/" % section)
     content = response.content.decode()
@@ -35,4 +68,3 @@ def reverse(*args, **kwargs):
             # raise NoReverseMatch only after testing final host
             if i == len(hosts)-1:
                 raise
-
